Tested Spi Dma Implementation. Works. Fixed GpioSettings return value that wasnt a reference. Added Dma hal and channels to spi hal for easy configuration . Made Nvic struct 32 bits wide always. Modified SpiPolled status and spot to utilize a variable of num data. Everything works great!

+++ STM/lld/Non_hal/NVIC/V1/nvic_lld.h
@@ -18,7 +18,7 @@ extern "C" {
 
 struct NvicHal
 {
-	const uint8_t num_irq; //number of interrupts to enable in for the peripheral
+	const uint32_t num_irq; //number of interrupts to enable in for the peripheral
 
 	const IRQn_Type *irq_number; //IRQn_Type from stm32 driver definitions header
 };
